+++ src/api/ecommerce/models/store.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 AccelByte Inc. All Rights Reserved.
+ * Copyright (c) 2021-2023 AccelByte Inc. All Rights Reserved.
  * This is licensed software from AccelByte Inc, for limitations
  * and restrictions contact your company contract manager.
  */
@@ -7,6 +7,10 @@
 import * as ioTs from "io-ts";
 import { DecodeError } from "../../types";
 
+/**
+ * @deprecated this is for old extension features thich still use io-ts.
+ * Please use `StoreInfo` from `@accelbyte/sdk` instead.
+ */
 export const Store = ioTs.intersection([
   ioTs.type({
     storeId: ioTs.string,
@@ -24,6 +28,11 @@ export const Store = ioTs.intersection([
     supportedRegions: ioTs.array(ioTs.string),
   }),
 ]);
+
+/**
+ * @deprecated this is for old extension features thich still use io-ts.
+ * Please use `StoreInfo` from `@accelbyte/sdk` instead.
+ */
 export type Store = ioTs.TypeOf<typeof Store>;
 
 export class StoreDecodeError extends DecodeError {}