+++ src/api/iam/role.ts
@@ -1,27 +1,20 @@
 /*
- * Copyright (c) 2021 AccelByte Inc. All Rights Reserved.
+ * Copyright (c) 2021-2023 AccelByte Inc. All Rights Reserved.
  * This is licensed software from AccelByte Inc, for limitations
  * and restrictions contact your company contract manager.
  */
 
-import { guardNetworkCall } from "../networkCallTypeguard";
-import { Network } from "../networkManager";
-import { Role, RoleDecodeError, RoleWithManagersResponse } from "./models/role";
+import { Validate } from "@accelbyte/sdk";
+import { AxiosInstance } from "axios";
+import { Role, RoleWithManagersResponse } from "./models/role";
 
-export const fetchRoles = (network: Network) => {
-  return guardNetworkCall(
+export const fetchRoles = (network: AxiosInstance) => {
+  return Validate.responseType<RoleWithManagersResponse>(
     () => network.get(`/iam/v3/admin/roles`),
-    RoleWithManagersResponse,
-    RoleDecodeError,
-    (error) => error
+    RoleWithManagersResponse
   );
 };
 
-export const fetchRole = (network: Network, { roleId }: { roleId: string }) => {
-  return guardNetworkCall(
-    () => network.get(`/iam/v4/admin/roles/${roleId}`),
-    Role,
-    RoleDecodeError,
-    (error) => error
-  );
+export const fetchRole = (network: AxiosInstance, { roleId }: { roleId: string }) => {
+  return Validate.responseType<Role>(() => network.get(`/iam/v4/admin/roles/${roleId}`), Role);
 };