+++ src/packages/Module/Submodule2/routes/Submodule2Routes.tsx
@@ -8,7 +8,7 @@ import React from "react";
 import { Route, Switch, useRouteMatch } from "react-router-dom";
 import { StoresPage } from "../pages/StoresPage";
 import { ItemsPage } from "../pages/ItemsPage";
-import { AllowedNamespaceGuard } from "../../../../components/AllowedNamespaceGuard/AllowedNamespaceGuard";
+import { AllowedNamespaceGuard } from "~/components/AllowedNamespaceGuard/AllowedNamespaceGuard";
 import subModule from "../submodule.json";
 
 export const Submodule2Routes = () => {