+++ src/typings.d.ts
@@ -1,0 +1,9 @@
+/*
+ *
+ *  Copyright (c) 2022 AccelByte Inc. All Rights Reserved.
+ *  This is licensed software from AccelByte Inc, for limitations
+ *  and restrictions contact your company contract manager.
+ *
+ */
+
+declare module "*.svg";