+++ CHANGELOG.md
@@ -2,6 +2,119 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.14](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.14%0D0.2.13) (2023-05-08)
+
+
+### Features
+
+* **extension-manifest-replacer:** add client name replacer ([72e4601](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/72e4601d17ae1d2ace049e3981ab3ee9f083f2c5)), closes [#57](https://accelbyte.atlassian.net/browse/57)
+* **extension:** upgrade extension website ([5dbd96b](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/5dbd96b69d8d0af22f5222506f42ea9dbdb49207)), closes [#56](https://accelbyte.atlassian.net/browse/56)
+
+### [0.2.13](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.13%0D0.2.12) (2023-04-10)
+
+### [0.2.12](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.12%0D0.2.11) (2023-03-13)
+
+
+### Bug Fixes
+
+* **extension:** add client name for sidebar menu ([b567f0e](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/b567f0e51196cd28426b8386e778066157a5a8d3)), closes [#53](https://accelbyte.atlassian.net/browse/53)
+* **user:** update AdminUser response model ([0308060](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/03080605a262cb6ab9cd2228c6bd5b292f2aaf3b)), closes [#54](https://accelbyte.atlassian.net/browse/54)
+
+### [0.2.11](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.11%0D0.2.10) (2023-02-13)
+
+### [0.2.10](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.10%0D0.2.9) (2023-01-16)
+
+
+### Features
+
+* **icons:** add basic icons & able to inject icons into AP core CEC-837 ([9ab707b](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/9ab707b3ea5527f51be01daa0e3ba9242c16f654)), closes [#50](https://accelbyte.atlassian.net/browse/50)
+
+
+### Bug Fixes
+
+* **extension-icon:** fix extension icon in index.html ([a29e535](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/a29e535c5315b4d92be54a198f23a0eceaccb7eb)), closes [#51](https://accelbyte.atlassian.net/browse/51)
+
+### [0.2.9](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.9%0D0.2.8) (2023-01-03)
+
+
+### Features
+
+* **extension-website:** fix error toast notification and add common validation error message ([6df7c1f](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/6df7c1fe2a1a15a0b2d84dbfb237b406c4fa21a5)), closes [#49](https://accelbyte.atlassian.net/browse/49)
+
+
+### Bug Fixes
+
+* **extension-website:** update extension website ([9068f9c](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/9068f9c41fe8017bf2b847310333e60b2091a960)), closes [#48](https://accelbyte.atlassian.net/browse/48)
+
+### [0.2.8](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.8%0D0.2.7) (2022-11-21)
+
+
+### Features
+
+* CWA-202 Add route prefix for AP ([747513e](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/747513ea417f02740de78a48f0a3fd9ddaa36373))
+* CWA-202 Replace prefix value placeholder with env var value ([7551322](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/755132255ad4fbe39ecca7cf6a4e9b0047660d57))
+
+
+### CI
+
+* CWA-202 Add JUSTICE_ADMINPORTAL_ROUTE_PREFIX to deployment ([9896c5d](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/9896c5dfec8a4c5d39921c574f78a1ad520e3d4d))
+
+### [0.2.7](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.7%0D0.2.6) (2022-10-24)
+
+### [0.2.6](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.6%0D0.2.5) (2022-08-01)
+
+### [0.2.5](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.5%0D0.2.4) (2022-07-18)
+
+
+### Features
+
+* **postmessage:** EN-112 add feature to share env JUSTICE_BASE_URL ([170a9e7](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/170a9e70800929e27cb2967ff415328208efb995)), closes [#34](https://accelbyte.atlassian.net/browse/34)
+* **sidebar-devmode:** EN-113 add dev mode sidebar ([9f0a64d](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/9f0a64dad6773ab3723ec02e9ba313f5f479b713)), closes [#33](https://accelbyte.atlassian.net/browse/33)
+
+### [0.2.4](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.4%0D0.2.3) (2022-07-04)
+
+### [0.2.3](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.3%0D0.2.2) (2022-03-01)
+
+### [0.2.2](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.2%0D0.2.1) (2022-02-14)
+
+
+### Features
+
+* **iframe:** use parent for downloading file ([4a5369f](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/4a5369f6278d8f330675b664cc467a2ad52b9130)), closes [#23](https://accelbyte.atlassian.net/browse/23)
+
+
+### Bug Fixes
+
+* **iframe:** enable toast notification to send a message in a form of JSX element ([69c8a28](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/69c8a28040f78a4f4ce1ab03b94103341b355126)), closes [#25](https://accelbyte.atlassian.net/browse/25)
+
+
+### Builds
+
+* update registry to default, get justice-js-common-utils from github ([a300748](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/a3007488e42e5d295181f011a4cfaba9cca4997a))
+
+### [0.2.1](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.1%0D0.2.0) (2021-11-22)
+
+
+### Features
+
+* **oauth:** Add refresh token when access admin portal extension inside admin portal ([ee67e34](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/ee67e34c97195c1ca36da8db1af1e33833e8933a))
+
+
+### Refactors
+
+* Module example follow standardization ([b1a1347](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/b1a134745a8e9dd79f81574f7f0855ab43dc5fa8))
+
+
+### Documentations
+
+* fix code block and add url pattern info and how it works ([f130d72](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/f130d72a29fdf76439b2bf82fea12f67ad18b216))
+* Update development guideline ([c7b56e8](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/c7b56e8462db87e857fe52c24797c8e3281e88b5))
+
+
+### Builds
+
+* CORE-5097 get justice-ui-library from github ([cd735d1](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/commits/cd735d1cf8d252d42f537546d52f0eab92cf8c46))
+
 ## [0.2.0](https://bitbucket.org/accelbyte/justice-adminportal-extension-website/branches/compare/0.2.0%0D0.1.2) (2021-10-25)